pyproject.toml
[project]
name = "doc-toc-and-summary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mistralai>=1.9.11",
    "openai>=2.6.1",
    "pandas>=2.3.3",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.7",
    "pandas-stubs>=2.3.2.250926",
    "types-requests>=2.32.4.20250913",
]
